requirements.txt
openai==1.3.0
semantic-kernel==0.4.0
pydantic==2.5.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiofiles==23.2.1
asyncio-mqtt==0.16.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
aiohttp==3.9.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dateutil==2.8.2
